+++ R/get_geoids.R
@@ -33,20 +33,32 @@
 #'   census will be returned.
 #'
 #' @export
-get_geoids <- function(geography, state = NULL, county = NULL, geoid = NULL,
-                       year = 2010, cache_tables = TRUE, key = NULL, ...) {
+get_geoids <- function(geography,
+                       state        = NULL,
+                       county       = NULL,
+                       geoid        = NULL,
+                       year         = 2010,
+                       cache_tables = TRUE,
+                       key          = NULL,
+                       ...) {
   
-  geography <- match.arg(geography,
-                         c("state", "county", "tract", "block group", "block"))
+  geography <-
+    match.arg(
+      geography,
+      c("state", "county", "tract", "block group", "block")
+    )
   
-  year      <- floor(year / 10) * 10
+  year <- floor(year / 10L) * 10L
   
-  if(!(year %in% c(1990, 2000, 2010))) {
+  if (!any(c(1990L, 2000L, 2010L) == year)) {
     stop("year must be between 1990 and 2019")
   }
   
-  variables <- stats::setNames(ifelse(year == 1990, "P0010001", "P001001"),
-                               paste0("census_", year, "_pop"))
+  variables <-
+    stats::setNames(
+      ifelse(year == 1990, "P0010001", "P001001"),
+      paste0("census_", year, "_pop")
+    )
   
   ref_area  <- validate_location(geoid, state, county, geography, ...)
   
@@ -54,19 +66,26 @@ get_geoids <- function(geography, state = NULL, county = NULL, geoid = NULL,
     list(
       geography    = geography,
       variables    = variables,
-      cache_tables = cache_tables,
+      cache_table  = cache_tables,
       key          = key,
       output       = "wide")
   
-  census_data <- call_tidycensus(fn           = tidycensus::get_decennial,
-                                 args         = args,
-                                 state_county = ref_area$state_county)
+  census_data <-
+    call_tidycensus(
+      fn           = tidycensus::get_decennial,
+      args         = args,
+      state_county = ref_area$state_county
+    )
   
-  if(!is.null(ref_area$geoid)) {
-    census_data <- filter_ref_area(data       = census_data,
-                                   geoid      = ref_area$geoid,
-                                   geo_length = ref_area$geo_length)
+  if (!is.null(ref_area$geoid)) {
+    census_data <-
+      filter_ref_area(
+        data       = census_data,
+        what       = "GEOID",
+        pattern    = ref_area$geoid,
+        geo_length = ref_area$geo_length
+      )
   }
   
-  return(census_data)
+  census_data
 }